Move messages-related classes to messages package
