+++ app/src/main/res/layout/fragment_messages.xml
@@ -5,7 +5,7 @@
     android:id="@+id/root_layout"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".ui.MessagesFragment">
+    tools:context=".ui.messages.MessagesFragment">
 
     <!-- TODO: make avatar circular -->
