+++ app/src/main/res/navigation/navigation_graph.xml
@@ -36,7 +36,7 @@
     </fragment>
     <fragment
         android:id="@+id/messagesFragment"
-        android:name="com.shapeapp.shape.ui.MessagesFragment"
+        android:name="com.shapeapp.shape.ui.messages.MessagesFragment"
         android:label="fragment_messages"
         tools:layout="@layout/fragment_messages">
         <action
